observlib/observlib/autoinstrumentation.py
from opentelemetry.instrumentation.threading import ThreadingInstrumentor
from opentelemetry.instrumentation.asyncio import AsyncioInstrumentor
from opentelemetry.instrumentation.logging import LoggingInstrumentor

LoggingInstrumentor().instrument(set_logging_format=True)
ThreadingInstrumentor().instrument()
AsyncioInstrumentor().instrument()
